requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-environ==0.4.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
httpie==2.1.0
idna==2.9
pkg-resources==0.0.0
psycopg2-binary==2.8.5
Pygments==2.6.1
PyJWT==1.7.1
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
